[goormlevel] Title: 운동 중독 플레이어, Time: 0.09 ms, Memory: 31.72 MB, Difficulty: 1 -BaekjoonHub

+++ goormlevel/195683/1. 운동 중독 플레이어/운동 중독 플레이어.java
@@ -7,7 +7,7 @@ class Main {
 		double answer = 0;
 	
 		
-		answer = Integer.parseInt(arr[0]) * (1 + Integer.parseInt(arr[1])/30.0);
+		answer = Double.parseDouble(arr[0]) * (1 + Double.parseDouble(arr[1])/30);
 		int result = (int) Math.floor(answer);
 		System.out.println(result);
